signals now handled for deleting an event; confirm dialog box for deleting; saveEventWidget doesn't close the screen when user clicked save and end time occurs before start time

+++ detailview.cpp
@@ -3,6 +3,7 @@
 #include <QLabel>
 #include <QPushButton>
 #include <QFormLayout>
+#include <QMessageBox>
 
 DetailView::DetailView(QRect &pageGeometry, Event_set &set, QWidget *parent) :
     RotaryView(pageGeometry, set, parent)
@@ -52,22 +53,32 @@ DetailView::DetailView(QRect &pageGeometry, Event_set &set, QWidget *parent) :
     formLayout->addWidget(deleteButton);
 
     connect(editButton, SIGNAL(clicked()), this, SLOT(slotEditClicked()));
-    //connect(deleteButton, SIGNAL(clicked()), this, SLOT(slotDeleteClicked()));
+    connect(deleteButton, SIGNAL(clicked()), this, SLOT(slotDeleteClicked()));
 }
 
 void DetailView::setCurrentEvent(Event* e)
 {
     currentEvent = e;
+    if(e == NULL)
+    {
+        title->setText("");
+        location->setText("");
+        description->setText("");
+        startTime->setText("");
+        endTime->setText("");
+    }
+    else
+    {
+        QDateTime qdt;
 
-    QDateTime qdt;
-
-    title->setText(e->getName());
-    location->setText(e->getLocation());
-    description->setText(e->getDescription());
-    startTime->setText(qdt.fromTime_t(e->getStartTime()).\
-                       toString(tr("ddd MMM, dd yyyy hh:mm")));
-    endTime->setText(qdt.fromTime_t(e->getEndTime()).\
-                     toString(tr("ddd MMM, dd yyyy hh:mm")));
+        title->setText(e->getName());
+        location->setText(e->getLocation());
+        description->setText(e->getDescription());
+        startTime->setText(qdt.fromTime_t(e->getStartTime()).\
+                           toString(tr("ddd MMM, dd yyyy hh:mm")));
+        endTime->setText(qdt.fromTime_t(e->getEndTime()).\
+                         toString(tr("ddd MMM, dd yyyy hh:mm")));
+    }
 }
 
 void DetailView::slotEditClicked()
@@ -78,10 +89,24 @@ void DetailView::slotEditClicked()
 
 void DetailView::slotDeleteClicked()
 {
+    if(currentEvent == NULL)
+        return;
+
+    // Ask the user to confirm
+    QMessageBox confirm(this);
+    confirm.setStandardButtons(QMessageBox::Ok | QMessageBox::Cancel);
+    confirm.setText(tr("<p>Are you sure you want to delete this event?</p>"));
+    confirm.show();
+    QPoint center = this->rect().center();
+    center.setX(center.x() - (confirm.width() / 2));
+    confirm.move(center);
+    int r = confirm.exec();
+    if(r == QMessageBox::Cancel)
+        return;
+
+    // Carry on and delete the event
     Event_set& set = this->getEventSet();
     set.deleteEvent(currentEvent);
-    //currentEvent = NULL;
-
-    //emit deleteEventClicked(currentEvent);
+    this->setCurrentEvent(NULL);
 }
